+++ src/main/java/com/example/repository/StatusRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.entity.Status;
 
-public interface StatusRepo extends JpaRepository<Status, Integer> {
+public interface StatusRepo extends JpaRepository<Status, Long> {
 
 }
