+++ src/main/java/com/example/service/StatusService.java
@@ -17,7 +17,6 @@ import com.example.entity.Product;
 import com.example.entity.Role;
 import com.example.entity.Status;
 import com.example.entity.Supplier;
-import com.example.entity.User;
 import com.example.exception.DatabaseConstraintException;
 import com.example.exception.DuplicateEmailException;
 import com.example.exception.ExceptionHandling;
@@ -27,14 +26,13 @@ import com.example.repository.ProductRepo;
 import com.example.repository.RoleRepo;
 import com.example.repository.StatusRepo;
 import com.example.repository.SupplierRepo;
-import com.example.repository.UserRepo;
 import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Service
 @Transactional
-public class StatusService extends PosService<Status, Integer> {
+public class StatusService extends PosService<Status, Long> {
 	private final StatusRepo repository;
 
 	public StatusService(StatusRepo repository) {
